# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DropFlow - Dropshipping Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/bulk-update:
    post:
      operationId: bulkUpdateOrders
      tags: [orders]
      summary: Bulk update status and/or tracking number on multiple orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderBulkUpdate"
      responses:
        "200":
          description: Bulk update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBulkUpdateResult"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order (status, tracking, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- SUPPLIER FINDER ----------
...
  /price-watch:
    get:
      operationId: listPriceWatches
      tags: [price-watch]
      summary: List all tracked competitor products
      responses:
        "200":
          description: List of price watches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceWatch"
    post:
      operationId: createPriceWatch
      tags: [price-watch]
      summary: Add a competitor product to watch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceWatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceWatch"
  /price-watch/{id}:
    delete:
      operationId: deletePriceWatch
      tags: [price-watch]
      summary: Remove a tracked product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /price-watch/{id}/snapshots:
    get:
      operationId: listPriceSnapshots
      tags: [price-watch]
      summary: Get price snapshot history for a tracked product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Snapshot history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceSnapshot"
    post:
      operationId: addPriceSnapshot
      tags: [price-watch]
      summary: Log a new price snapshot for a tracked product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceSnapshotInput"
      responses:
        "201":
          description: Snapshot recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceSnapshot"
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, status, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        productId:
          type: ["integer", "null"]
        productName:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        quantity:
          type: integer
        costPrice:
          type: ["number", "null"]
        sellPrice:
          type: ["number", "null"]
        profit:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, placed, shipped, delivered, cancelled]
        trackingNumber:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        placedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, quantity]
      properties:
        orderNumber:
          type: string
        productId:
          type: integer
        supplierId:
          type: integer
        customerName:
          type: string
          minLength: 1
        customerEmail:
          type: string
        quantity:
          type: integer
        costPrice:
          type: number
        sellPrice:
          type: number
        status:
          type: string
          enum: [pending, placed, shipped, delivered, cancelled]
        shippingAddress:
          type: string
        trackingNumber:
          type: string
    OrderUpdate:
      type: object
      properties:
        productId:
          type: integer
        supplierId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        quantity:
          type: integer
        costPrice:
          type: number
        sellPrice:
          type: number
        status:
          type: string
          enum: [pending, placed, shipped, delivered, cancelled]
        trackingNumber:
          type: string
        shippingAddress:
          type: string
        placedAt:
          type: string

    # ---- Dashboard Analytics ----
...
    OrderBulkUpdate:
      type: object
      required: [orderIds]
      properties:
        orderIds:
          type: array
          items:
            type: integer
          minItems: 1
        status:
          type: string
          enum: [pending, placed, shipped, delivered, cancelled]
        trackingNumber:
          type: string
    OrderBulkUpdateResult:
      type: object
      required: [updatedCount, orders]
      properties:
        updatedCount:
          type: integer
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"

    # ---- Supplier Finder ----
...
    PriceWatchInput:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string
          minLength: 1
        myPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    PriceWatch:
      type: object
      required: [id, name, url, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        myPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        latestPrice:
          type: ["number", "null"]
        latestRecordedAt:
          type: ["string", "null"]
        snapshotCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    PriceSnapshotInput:
      type: object
      required: [price]
      properties:
        price:
          type: number
          minimum: 0
        note:
          type: ["string", "null"]
    PriceSnapshot:
      type: object
      required: [id, watchId, price, recordedAt]
      properties:
        id:
          type: integer
        watchId:
          type: integer
        price:
          type: number
        note:
          type: ["string", "null"]
        recordedAt:
          type: string

    # ---- Dashboard ----
...